services/data-ingestion/requirements.txt
paho-mqtt
pymysql
cryptography
